blaze(nn): fp16 NHWC act store, fp32 accumulate
Conv graph and d_obs materialize fp16. Master weights, dense hidden, Adam, and GAE stay fp32. Cam FC upcasts conv2 Y.

+++ blaze/nn/cuda_conv_graph.h
@@ -1,9 +1,10 @@
-/* Shape-agnostic cuDNN-graph conv net. NHWC acts, KRSC filters, TF32.
+/* Shape-agnostic cuDNN-graph conv net. NHWC/KRSC fp16 store, fp32 accumulate.
  * One spec per layer. FWD = conv + bias + ReLU. See cuda_fable_contract.h. */
 #pragma once
 
 #include "cuda_ws.h"
 
+#include <cuda_fp16.h>
 #include <cudnn.h>
 #include <stddef.h>
 
@@ -38,17 +39,17 @@ int nn_conv_net_c_out(const NnConvNet *net, int layer);
 int nn_conv_net_h_out(const NnConvNet *net, int layer);
 int nn_conv_net_w_out(const NnConvNet *net, int layer);
 
-/* x,y NHWC fp32. w KRSC fp32. bias [c_out]. Layer 0 x uses padded C. */
-int nn_conv_net_fwd(NnConvNet *net, int layer, const float *x, const float *w,
-                    const float *bias, float *y);
+/* x,y NHWC fp16. w KRSC fp16. bias [c_out] fp16. Layer 0 x uses padded C. */
+int nn_conv_net_fwd(NnConvNet *net, int layer, const __half *x, const __half *w,
+                    const __half *bias, __half *y);
 
-/* dpre = dy after ReLU bwd. beta: 0 overwrite, 1 accumulate into dw. */
-int nn_conv_net_wgrad(NnConvNet *net, int layer, const float *x,
-                      const float *dpre, float *dw, float beta);
+/* dpre = dy after ReLU bwd (fp16). dw KRSC fp16. beta: 0 overwrite, 1 acc. */
+int nn_conv_net_wgrad(NnConvNet *net, int layer, const __half *x,
+                      const __half *dpre, __half *dw, float beta);
 
 /* Skip for layer 0 (returns 0, no-op). */
-int nn_conv_net_dgrad(NnConvNet *net, int layer, const float *w,
-                      const float *dpre, float *dx);
+int nn_conv_net_dgrad(NnConvNet *net, int layer, const __half *w,
+                      const __half *dpre, __half *dx);
 
 #ifdef __cplusplus
 }
